+++ src/main/java/org/folio/search/service/browse/CallNumberBrowseQueryProvider.java
@@ -18,6 +18,7 @@ import org.folio.search.configuration.properties.SearchQueryConfigurationPropert
 import org.folio.search.model.service.BrowseContext;
 import org.folio.search.model.service.BrowseRequest;
 import org.folio.search.service.metadata.SearchFieldProvider;
+import org.folio.search.service.setter.item.ItemCallNumberProcessor;
 import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.script.Script;
 import org.opensearch.search.builder.SearchSourceBuilder;
@@ -33,7 +34,7 @@ public class CallNumberBrowseQueryProvider {
   private static final int MIN_QUERY_SIZE = 25;
 
   private final SearchFieldProvider searchFieldProvider;
-  private final CallNumberTermConverter callNumberTermConverter;
+  private final ItemCallNumberProcessor callNumberProcessor;
   private final SearchQueryConfigurationProperties queryConfiguration;
   private final CallNumberBrowseRangeService callNumberBrowseRangeService;
 
@@ -65,7 +66,7 @@ public class CallNumberBrowseQueryProvider {
 
   private QueryBuilder getQuery(BrowseContext ctx, String tenantId, int size, boolean isBrowsingForward) {
     var anchor = ctx.getAnchor();
-    var callNumberAsLong = callNumberTermConverter.convert(anchor);
+    var callNumberAsLong = callNumberProcessor.getCallNumberAsLong(anchor);
     var rangeQuery = rangeQuery(CALL_NUMBER_RANGE_FIELD);
     rangeQuery = isBrowsingForward ? rangeQuery.gte(callNumberAsLong) : rangeQuery.lte(callNumberAsLong);